updated urls, fixed some errors

+++ game-court-application/src/main/java/dev/sukanya/gamecourtbooking/dto/cost/CostType.java
@@ -4,5 +4,5 @@ import dev.sukanya.gamecourtbooking.dto.timeslot.TimeSlotDTO;
 
 public interface CostType {
 
-    long costCalculationAlgorithm(long specialCourtChargePerHour, TimeSlotDTO selectedTimeSlot);
+    int costCalculationAlgorithm(int specialCourtChargePerHour, TimeSlotDTO selectedTimeSlot);
 }
